Add definition of the pmix_coord_t struct
Provide a structure by which we can describe the network coordinate of a
process. This is done using a new pmix_coord_t data type with supporting
attributes to identify the associated coordinate system. In addition, a
new attribute for getting/querying the network coordinates of a
specified process is provided.

Signed-off-by: Ralph Castain <rhc@pmix.org>