Clarify regular expression generator output
The PMIx_generate_regex and PMIx_generate_ppn functions are intended
solely to prepare input for PMIx_server_register_nspace. The rationale
is that someone might need to only run those generators once per nspace,
transmitting the output to remote nodes for input to register_nspace.

This modification simply clarifies that the output from those generator
functions may not be a NULL-terminated string, but instead could be a
byte array of arbitrary binary content.

Signed-off-by: Ralph Castain <rhc@pmix.org>