Fix signature of blocking PMIx_Allocate_request
The blocking form of PMIx_Allocate_request was added in v3.1, but
unfortunately the signature didn't include the necessary parameters to
return the results of the request. This PR corrects that oversight.

Thanks to the Spark team for pointing it out.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit d843065896281e5f100a547a872e92de3cbca19f)