Cleanup some blocking functions
Some of the blocking functions were missing necessary parameters to
return the results of their operation. In some casees, no blocking
form of an operation was provided - where possible, we provide a
blocking form by simply looking for a NULL callback function. We
return the result of the operation in the status code.

Signed-off-by: Ralph Castain <rhc@pmix.org>