Add/update the fabric support chapter
Consolidate the scheduler and coordinate chapters into a single chapter
on fabric support. Pull all fabric-related APIs, data types, constants,
and attributes into the new chapter. Update the definition of the
pmix_fabric_t object to be a little more abstract. Provide a new API for
requesting an update to the fabric object for systems that need to poll
for updates rather than use the event notification method.

Signed-off-by: Ralph Castain <rhc@pmix.org>