Cleanup registration info description
Remove the reference to the pmix.org URL in the PMIx_Get description and
replace it with a reference to the PMIx_server_register_nspace function
where the required/optional information is defined along with a
statement on how to access the various provided info. Access methods are
a part of the Standard and are not implentation-specific.

Rewrite the PMIx_server_register_nspace attribute section to more
clearly delineate what is required vs optional, and to split out the
session, job, app, proc, and node-level sections.

Deprecate PMIX_PROC_DATA in favor of PMIX_PROC_INFO_ARRAY for symmetry
so it corresponds to its session, job, app, and node brethren.

Signed-off-by: Ralph Castain <rhc@pmix.org>