Extend topology support
Users and applications would like a scalable way to retrieve the local
topology, preferably via a shared memory region from the local
persistent daemon. This alleviates the discovery time problem for large
ppn scenarios as well as the memory footprint penalty associated with
every proc carrying its own topology description.

Define a new pmix_topology_t abstraction that includes the source of the
topology description (e.g., "hwloc") and a void* pointer to the
description itself. Add a PMIX_TOPOLOGY2 attribute to use that value
in place of hwloc_topology_t to provide more of an abstraction.
Deprecate PMIX_TOPOLOGY. Server library is responsible for cleaning up
at finalize.

When PMIX_TOPOLOGY or PMIX_TOPOLOGY2 is passed to a server during PMIx_server_init, have
the server be responsible for setting up whatever support is available
(depending on topology implementation) and passing the required keys
down to the client. This removes all the current PMIX_* topology keys as
those become hidden inside the PMIx implementation - so deprecate them
in this version.

Add a new PMIx_Load_topology API by which the user/app can retrieve the
topology description for their subsequent use. Strongly warn that the
returned pointer is to be treated as "read only" and that attempts to
modify it will return an error. Client library is responsible for
cleaning up at finalize.

Do a little cleanup of long lines so diff is easier to read. Fix an
error in naming convention - it should be PMIx_Get_relative_locality
with a capital-G as it is a client-facing function.

Signed-off-by: Ralph Castain <rhc@pmix.org>
Co-authored-by: Josh Hursey <jhursey@us.ibm.com>