Modify process set definitions
Per feedback, allow processes to belong to more than one process set at
a time. Clarify that process sets are defined by the host environment as
opposed to process groups that are defined by the application. Add APIs
for the host to inform the servers of changes to process sets, and event
statuses to allow the servers to communicate those changes to their
client procs.

Signed-off-by: Ralph Castain <rhc@pmix.org>