Modify process set definitions
Per feedback, allow processes to belong to more than one process set at
a time. Clarify that process sets are defined by the host environment as
opposed to process groups that are defined by the application. Add APIs
for the host to inform the servers of changes to process sets, and event
statuses to allow the servers to communicate those changes to their
client procs.

Clarify that process set gets deleted upon removal of last member

Do not include process groups in queries about process sets. Add
group-specific attributes for that purpose.

Change return value of query for pset names to pmix_data_array_t and
return an array of char* to avoid defining a delimiting character

Move PMIX_PSET_NAMES to the process set chapter and update its
description.

Add a corresponding PMIX_GROUP_NAMES.

Clarify that there is no linkage between a process set and process group
of the same name.

Signed-off-by: Ralph Castain <rhc@pmix.org>