Further cleanup of the fabric and topology support
Rename PMIx_Fabric_update_distances to PMIx_Fabric_compute_distances and
provide parameters to support computing of distances for given
topologies. This allows support for hetero situations.

Change PMIx_Get_cpuset so it actually returns the bitmap of the current
process or thread, as specified.

Add PMIx_Parse_cpuset_string to convert a string to an actual bitmap.

Cleanup the distance computation APIs
Distances can be computed for many device types, not just fabric, so
rename the API accordingly and move it to the process locality area
since it is locality-related. Define a data type by which the caller can
specify the type(s) of device(s) they want distances to, along with the
topology and cpuset to use.

Signed-off-by: Ralph Castain <rhc@pmix.org>