Final touchups to PMIx_Get and PMIx_server_init attributes
Provide the ability to request pointers to values, filling in of a
provided value struct address, as well as the default return of a newly
malloc'd struct. Clarify that "success" means the data was returned in
the form requested - do not return it as something different.

Remove the "slot" definition per WG discussion

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit d86b03136103a6389225a0e0286add2b39c7fe94)