Assign specific values to PMIx constants
The Standard currently only defines the name of PMIx constants - it does
not assign them a value, thus leaving the value to each implementation.
This creates a problem for vendors seeking to distribute
implementation-agnostic products as they cannot know the meaning of a
particular value returned by a function call.

Implementation-agnostic products (e.g., debuggers) are typically compiled
against one implementation, but are designed to work against whatever
implementation is installed on the target system. This is done by using
the combination of dlopen and dlsym to load the library and access its
public APIs.

Unfortunately, the value of the constants in the code is determined at
compile time - it isn't possible to load them via dlopen. Thus, the
resulting binary is "locked" to the implementation against which it was
compiled, thereby forcing the vendor to distribute
implementation-specific copies of their product.

The solution to the problem is to have the Standard define the actual
value of each constant and not just its name. Doing this forces all
implementations to use the same value, thus allowing for
implementation-agnostic products to be distributed.

Signed-off-by: Ralph Castain <rhc@pmix.org>
Signed-off-by: John DelSignore <JDelSignore@perforce.com>