Define tool in terms chapter
I don't recall why this was never defined.  We may need to think about this.
Certainly I could have just said, "A tool is a process that calls PMIx_tool_init()"
but I wanted to carefully contrast it with a client so I defined it in similar
language even though it uses the gramatically wasteful approach of
X is either a FOO or ~FOO and BAR.  We could have said a client is a process that
calls PMIx_Init(), but we didn't because how it is started is important in some way.

Signed-off-by:  dsolt@us.ibm.com