Add some missing attributes and one API for server support
There are a few attributes servers need in order to support
singleton operations as well as user-specified time limits
on job and spawn operations. Current server module functions
do not include the ability for the server library to pass
a local status from a collective operation (i.e., when the
server library aggregates all local participation before
calling the host)

Add a few job-related error constants for clear reporting
to the user.

Add a useful variant of a data management function.

Signed-off-by: Ralph Castain <rhc@pmix.org>