Move synchronization function to its own chapter
We moved the synchronization material out of Sync_Access into its own
chapter. The new Chap_API_Sharing_Basics.tex is identical to the old
Chap_API_Sync_Access chapter except the introduction changed and the
synchronization API's are deleted.  The Sharing_Basics introduction
is very minimal right now, but we have additional changes coming to
populate this chapter with an thorough explanation of the get/put
routines.

Signed-off-by: dsolt@us.ibm.com