Move NonReserved_Keys (going away) to SyncAccess
The chapter on NonReserved_Keys is going away.  The final location
is a new chapter called SharingBasics that goes over the basics of
how applications share data.  We then present the idea that there is
some "already shared data" in the form of reserved keys in that chapter.
This can all get collapsed into a simpler commit when we merge, but
doing it this way helps with viewing how text changed.

Signed-off-by: dsolt@us.ibm.com