Provide access to resource usage for processes and nodes
Operating systems typically maintain a running measure of resource
utilization by active processes. This includes metrics on CPU
utilization, disk accesses, memory size, and network activity.
Define a set of attributes by which these these metrics can be
requested and returned.

Attributes are used as a means of providing for later extension
to include a broader range of metrics.

Signed-off-by: Ralph Castain <rhc@pmix.org>