Replace the "log" server module fn with a "log2"
The current "log" server module function was defined with
a void return. This contrasts with most of the other module
functions which characteristically return a pmix_status_t,
thereby supporting a return status that indicates if the
function is not supported or has been atomically completed.

Add a replacement "log2" function that is identical in signature
except for returning a pmix_status_t instead of void. Deprecate
the current "log" function.

Signed-off-by: Ralph Castain <rhc@pmix.org>

Provide better description for PMIx_Log arguments

Expand on the use of the "data" vs "directives" attribute
arrays, providing a couple of examples to help clarify
their use.

Signed-off-by: Ralph Castain <rhc@pmix.org>