Errata: fix a couple of spots in server upcall functions
Server upcall functions were required to return prior to executing
the callback function. This was a limitation of the implementation
and not a requirement of the Standard - the statements should
never have been included here.

FWIW: the limitation has subsequently been removed from the
reference implementation.

In the pmix_server_iof_fn_t, clarify that an error status can be
returned either directly by the function or in the callback function.
It does not require that the host execute the callback to return
an error, as was implied by the current language.

Signed-off-by: Ralph Castain <rhc@pmix.org>