Convert the PMIX_VALUE_GET_NUMBER macro to a function
Converting this macro to a PMIx_Value_get_number function requires
a signature change - it is no longer possible to pass the C datatype
for casting the result. Instead, we now pass the PMIx datatype that
corresponds to the desired C casting. We also extend the function
to return a new PMIX_ERR_LOST_PRECISION status when the casting
can potentially cause a loss of precision - e.g., when changing from
a "float" to an "int" - and a new PMIX_ERR_CHANGE_SIGN status
when the casting would cause a change in sign - e.g., when changing
from a negative integer to a "size_t".

Signed-off-by: Ralph Castain <rhc@pmix.org>